refactor: extract duplicated splitText into shared gateway utility
